+++ internal/luasandbox/service.go
@@ -2,7 +2,6 @@ package luasandbox
 
 import (
 	"context"
-	"time"
 
 	lua "github.com/yuin/gopher-lua"
 
@@ -21,57 +20,14 @@ func newService(
 	}
 }
 
-type RunOptions struct {
-	Source  string
+type CreateOptions struct {
 	Modules map[string]lua.LGFunction
-	Timeout time.Duration
 }
 
-const DefaultTimeout = time.Millisecond * 200
-
-func (s *Service) Run(ctx context.Context, opts RunOptions) (err error) {
-	ctx, endObservation := s.operations.run.With(ctx, &err, observation.Args{})
+func (s *Service) CreateSandbox(ctx context.Context, opts CreateOptions) (_ *Sandbox, err error) {
+	ctx, endObservation := s.operations.createSandbox.With(ctx, &err, observation.Args{})
 	defer endObservation(1, observation.Args{})
 
-	// Configure execution timeout
-	if opts.Timeout == 0 {
-		opts.Timeout = DefaultTimeout
-	}
-	ctx, cancel := context.WithTimeout(ctx, opts.Timeout)
-	defer cancel()
-
-	// Create a new Lua virtual machine
-	state := s.creatNewStateWithModules(ctx, opts.Modules)
-	defer state.Close()
-
-	// Execute supplied program
-	if err := state.DoString(opts.Source); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Modified verison of the libs available from gopher-lua@linit.go:luaLibs
-var builtinLibs = []struct {
-	libName string
-	libFunc lua.LGFunction
-}{
-	{lua.BaseLibName, lua.OpenBase},
-	{lua.ChannelLibName, lua.OpenChannel},
-	{lua.CoroutineLibName, lua.OpenCoroutine},
-	{lua.DebugLibName, lua.OpenDebug},
-	{lua.LoadLibName, lua.OpenPackage},
-	{lua.MathLibName, lua.OpenMath},
-	{lua.StringLibName, lua.OpenString},
-	{lua.TabLibName, lua.OpenTable},
-
-	// Explicitly omitted
-	// {lua.IoLibName,lua. OpenIo},
-	// {lua.OsLibName, lua.OpenOs},
-}
-
-func (s *Service) creatNewStateWithModules(ctx context.Context, modules map[string]lua.LGFunction) *lua.LState {
 	state := lua.NewState(lua.Options{
 		// Do not open libraries implicitly
 		SkipOpenLibs: true,
@@ -85,10 +41,12 @@ func (s *Service) creatNewStateWithModules(ctx context.Context, modules map[stri
 	}
 
 	// Preload caller-supplied modules
-	for name, loader := range modules {
+	for name, loader := range opts.Modules {
 		state.PreloadModule(name, loader)
 	}
 
-	state.SetContext(ctx)
-	return state
+	return &Sandbox{
+		state:      state,
+		operations: s.operations,
+	}, nil
 }
